feat(movie): add 3 admin endpoints in MovieController and everything they needed (handlers, validators, specification etc): - POST   /admin/movies/media - DELETE /admin/movies/{id:guid} - DELETE /admin/movies/{movieId:guid}/media/{mediaId:guid}
Also:
- tiny logic bug fix in Movie domain entity method "AddMedia".
- replaced hardcoded AsNoTracking() with conditional check based on spec.AsNoTracking flag in "BaseRepository" (now applies AsNoTracking only if spec.AsNoTracking == true)
- tiny fix of the duration field in "MovieConfiguration" ==> (v => (int)v.TotalSeconds) and created migration for that
- added controllers registration and swagger in "Program.cs"

+++ CinemaAura/AbsoluteCinema.Infrastructure/Persistence/Configurations/MovieConfiguration.cs
@@ -47,7 +47,7 @@ public class MovieConfiguration : IEntityTypeConfiguration<Movie>
         builder.Property(x => x.Duration)
             .HasColumnName("duration")
             .HasConversion(
-                v => v.TotalSeconds,
+                v => (int)v.TotalSeconds,
                 v => TimeSpan.FromSeconds(v))
             .HasColumnType("integer");
